# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hacker Dashboard API",
    "description": "임의 투자 데이터 스키마에서 자동 분석 뷰를 생성하는 금융 대시보드 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analyze": {
      "post": {
        "tags": [
          "analyze"
        ],
        "summary": "Analyze",
        "description": "임의 투자 데이터를 받아 Router → Analyzer → 3단 게이트 파이프라인을 실행한다.\n\n- `data` 만 주어진 경우: 그대로 분석\n- `symbol` 이 주어진 경우: market 어댑터로 OHLC 프리패치 후 주입\n- `context.ohlc` 가 주어진 경우: 그대로 rows 로 변환해 주입\n- `include_portfolio_context=True` 인 경우: DB에서 holdings를 조회해 AgentState에 주입",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/csv": {
      "post": {
        "tags": [
          "analyze"
        ],
        "summary": "Analyze Csv",
        "description": "CSV 파일 업로드 → 자동 분석.\n\n- content-type: text/csv 또는 application/vnd.ms-excel\n- 파일 크기: ≤ 10MB (초과 시 413)\n- 인코딩: utf-8 / cp949 자동 탐지\n- 행 수 > 300 이면 모델을 opus 로 자동 승급\n- 결과는 파일 해시 + user_note 기반 캐시 키로 5분 캐싱",
        "operationId": "analyze_csv_analyze_csv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_csv_analyze_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeContext": {
        "properties": {
          "ohlc": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OhlcBar"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ohlc",
            "description": "서버 또는 클라이언트가 제공한 OHLC 시계열"
          }
        },
        "type": "object",
        "title": "AnalyzeContext",
        "description": "선택적 부가 컨텍스트. 현재는 OHLC 프리패치 결과만."
      },
      "AnalyzeMeta": {
        "properties": {
          "asset_class": {
            "type": "string",
            "title": "Asset Class",
            "description": "Router 가 결정한 자산군"
          },
          "router_reason": {
            "type": "string",
            "title": "Router Reason",
            "description": "Router 가 선택한 이유 (한 문장)"
          },
          "gates": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Gates",
            "description": "각 게이트의 최종 상태"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "그래프 실행 소요 시간(ms)"
          },
          "analyzer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzer Name",
            "description": "어떤 analyzer 가 실행되었는지 (stock | crypto | fx | ...)"
          },
          "evidence_snippets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Snippets",
            "description": "critique 통과 후 확정된 근거 인용 3~5개"
          },
          "cache": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CacheMetrics"
              },
              {
                "type": "null"
              }
            ],
            "description": "직전 요청의 프롬프트 캐시/토큰 사용량"
          }
        },
        "type": "object",
        "required": [
          "asset_class",
          "router_reason",
          "gates"
        ],
        "title": "AnalyzeMeta"
      },
      "AnalyzeRequest": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data",
            "description": "분석할 원시 데이터 rows. symbol 만으로도 분석할 경우 빈 리스트 허용."
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "자연어 분석 지시 (선택)"
          },
          "asset_class_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Class Hint",
            "description": "Router 힌트: stock | crypto | fx | macro | auto"
          },
          "symbol": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SymbolRef"
              },
              {
                "type": "null"
              }
            ],
            "description": "심볼 분석 요청. 존재 시 서버가 market 어댑터로 OHLC 를 프리패치한다."
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeContext"
              },
              {
                "type": "null"
              }
            ],
            "description": "클라이언트가 선계산해 넘기는 OHLC 등 부가 컨텍스트"
          },
          "include_portfolio_context": {
            "type": "boolean",
            "title": "Include Portfolio Context",
            "description": "True면 서버가 DB에서 holdings를 조회해 AgentState에 주입. LLM이 개인화된 분석 수행.",
            "default": false
          }
        },
        "type": "object",
        "title": "AnalyzeRequest",
        "description": "임의 투자 데이터를 받아 자동 분석을 요청한다."
      },
      "AnalyzeResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "추적용 UUID"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ok | error | degraded"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "Analyzer 출력 (자산군별 상이)"
          },
          "meta": {
            "$ref": "#/components/schemas/AnalyzeMeta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status",
          "meta"
        ],
        "title": "AnalyzeResponse"
      },
      "Body_analyze_csv_analyze_csv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "user_note": {
            "type": "string",
            "title": "User Note",
            "default": ""
          },
          "symbol_hint": {
            "type": "string",
            "title": "Symbol Hint",
            "default": ""
          },
          "include_portfolio_context": {
            "type": "boolean",
            "title": "Include Portfolio Context",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_csv_analyze_csv_post"
      },
      "CacheMetrics": {
        "properties": {
          "read_tokens": {
            "type": "integer",
            "title": "Read Tokens",
            "description": "프롬프트 캐시 히트 토큰",
            "default": 0
          },
          "creation_tokens": {
            "type": "integer",
            "title": "Creation Tokens",
            "description": "캐시 생성(기록) 토큰",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          }
        },
        "type": "object",
        "title": "CacheMetrics"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OhlcBar": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts",
            "description": "ISO-8601 UTC"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "OhlcBar",
        "description": "OHLC 캔들 하나"
      },
...
      "SymbolRef": {
        "properties": {
          "market": {
            "type": "string",
            "title": "Market",
            "description": "upbit | binance | yahoo | naver_kr"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "티커/코드. 예: KRW-BTC, AAPL"
          }
        },
        "type": "object",
        "required": [
          "market",
          "code"
        ],
        "title": "SymbolRef",
        "description": "업로드 없이 심볼만으로 분석할 때 사용하는 참조."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}